tools/image_extractor/requirements.txt
# Image Extractor API Dependencies
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
requests>=2.31.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
ultralytics>=8.0.0
google-cloud-bigquery>=3.13.0
